[BE-1237] Fixed a first batch of compiler warnings (#4164)
* fixed compiler warnings for join

* fixed errors in parquet_write.cpp

* fixed most group by compiler warnings

* fixed import bodo

* removed unused file

* fix unused functions in .h files

* removed extern C

* remove slow test for cpp change

* Fixed groupby errors

* fixed virtual issue in ArrowDataframeReader

* fixed import issues, down to just 1 unused function

* Fixed the forward reference issue

* Fixed another forward reference issue

* fixed _csv_json_reader.cpp

* added separate util functions

* remove incorrect static

* fixed the macro names

* fixed tracing on CI

* revert the tracing back to size_t

* fixed the variable name

* fixed loop vars

* switched size_t to uint64_t in several places

* added signatures at the top

* made sahil's required change

* added Sahil's name change

* Made Hadia's changes

+++ bodo/libs/_meminfo.h
@@ -142,7 +142,7 @@ inline size_t NRT_MemSys_get_stats_mi_alloc() { return TheMSys.stats_mi_alloc; }
 inline size_t NRT_MemSys_get_stats_mi_free() { return TheMSys.stats_mi_free; }
 
 struct MemInfo {
-    size_t refct;
+    int64_t refct;
     NRT_dtor_function dtor;
     void *dtor_info;
     void *data;
@@ -202,7 +202,7 @@ inline void NRT_MemInfo_call_dtor(NRT_MemInfo *mi) {
 inline void *NRT_Allocate(size_t size) {
     void *ptr = TheMSys.allocator.malloc(size);
     if (!ptr) {
-        std::cerr << "bad alloc: possible Out of Memory error\n"; 
+        std::cerr << "bad alloc: possible Out of Memory error\n";
         exit(9);
     }
 #ifdef BODO_DEBUG
@@ -309,7 +309,8 @@ inline NRT_MemInfo *NRT_MemInfo_alloc_safe_aligned(size_t size,
     std::cerr << "NRT_MemInfo_alloc_safe_aligned " << data << " " << size
               << "\n";
 #endif
-    NRT_MemInfo_init(mi, data, size, nrt_internal_dtor_safe, (void *)size, NULL);
+    NRT_MemInfo_init(mi, data, size, nrt_internal_dtor_safe, (void *)size,
+                     NULL);
     return mi;
 }
